+++ modules/apps/src/test/java/com/axway/apim/cli/AppCLIOptionsTest.java
@@ -27,6 +27,7 @@ public class AppCLIOptionsTest {
 		// Validate App-Import parameters
 		Assert.assertEquals(params.getConfig(), "myAppConfig.json");
 		Assert.assertEquals(params.getStageConfig(), "myStageConfig.json");
+		Assert.assertNotNull(params.getProperties(), "Properties should never be null. They must be created as a base or per stage.");
 	}
 	
 	@Test
@@ -51,6 +52,7 @@ public class AppCLIOptionsTest {
 		Assert.assertEquals(params.getCreatedBy(), "Tom");
 		Assert.assertEquals(params.getCredential(), "*9877979779*");
 		Assert.assertEquals(params.getRedirectUrl(), "*localhost*");
+		Assert.assertNotNull(params.getProperties(), "Properties should never be null. They must be created as a base or per stage.");
 	}
 
 }